requirements.txt
sentence-transformers
peft
transformers
datasets
evaluate
rouge_score
nltk
torch
numpy
Pillow
tqdm
